[kernel][ktrace] Revamp static registration, add "ktrace_ready" probe
Change the static registration scheme so that the "orphan" section
created is read-only.  Writable "orphan" sections do not come out
right with LLD and we must avoid them.

Report an event for "tracing is all set up now".  This also serves
to ensure that there will be at least one static probe entry so that
registration is actually tested.  Since __{start,stop}_ktrace_probe
symbols will always be defined by the linker, we don't need to use
weak symbols, which force the compiler to use a GOT slot under PIC.

Along the way this makes the ktrace_probeN macros C++-compatible
and ;-eating.

Change-Id: I30c5aac0af1c5822e4aff3fedaac5dd961b97dfd