netconnector: add MdnsService.AddResponder
MdnsService.AddResponder is a more versatile way to publish a service
instance (compared to MdnsService.PublishServiceInstance). All
announcements and query responses are filtered through the client,
and service subtypes are supported. The client has the option of
announcing subtypes, if desired. In addition, ttls for PTR, SRV and
TXT records can be specified by the client.

Change-Id: I65d6298f0f6695a774a6fc1404fa73da0a09683f