[hypervisor][scripts] Add bootstrap-debian.sh
Script to create a debian sysroot image that is bootable as a
hypervisor guest.

The intention is that the filesystem will be placed on a USB drive and
the zircon block device file (ex /dev/class/block/XYZ) can be provided
to the guest as the block file.

The next step is to create scripts to setup an EFI partition with a
grub bootloader that can either boot linux (to run benchmarks on bare
metal or as a KVM guest) or chainload gigaboot to load zircon.

Change-Id: Ic3a3106420c67fdd80fc0b22de78ed06cb6b1abf