[kernel][arm64] Remove bogus assert from linker script
This assert is having no effect.  In fact .text.el2 is larger than a
page, and this is not detecting it.  To detect this in the linker
script requires rearranging how the section is emitted, which is
hard given its name (which matches .text* earlier).  But since the
invariant does not actually hold right now, actually fixing it would
break the build.

A simpler approach to enforce this is to just put ".org 0x1000" at
the end of el2.S.  That would pad it out to a full page if it were
less, and also give an error if it's more (as it is now).

Change-Id: I4b3cb0c548ecb42f1108195c4798c6f412af208e