[utest][x86] Disable segment selector test that hangs under non-KVM QEMU
This test hangs under some (but not all) versions of non-KVM QEMU,
because these versions don't implement the part of the IRET
instruction's behaviour that the test tests for.

I'm leaving the code in place so that the test can be run manually,
because it is useful for demonstrating IRET's behaviour.

ZX-1240

Change-Id: I6ea88e027a05c46ecd06ce56d8d0a93d037f0db7