[resolver] Add ModuleManifestRepository lib
This exposes the contents of /system/data/module_manifest_repository as
a stream of "verb implementation" structs. This utilizes
fdio_watch_directory(), which happens to be an easy way to both iterate
through the contents of a directory, as well as get notified of new
additions.

MW-273 #done
MW-274 #done

Change-Id: I222cfee9f5bde81cd07b94f027db9bc6d664771d