[media] Route get_min_delay call to audio_server, set timeline asynchronously
This makes two changes to the audio client library behavior to make
playback behavior more consistent. First, it routes the query for the
minimum theoretical stream delay to the audio_server. The call isn't
routed all the way into the output device yet, but this has the
practical side effect of 'warming up' the audio server in response to
the call. Second, this adds a fire-and-forget method for setting the
currently applicable timeline. This is useful for when the caller wants
to set the timeline at some point in the future but not wait for that
to happen or be woken up when it does.

Change-Id: I520527376aba332ecfe5f89b0f8b5b693c96cb55