[zxcrypt] Add Superblock class
This CL adds Superblock, which encapsulates how the cryptographic
metadata is stored and retrieved by zxcrypt.  I/O performance is not as
critical a concern here, as the superblock is generally only involved in
'lifecycle' events of a zxcrypt device, that is formatting, binding,
rekeying, and shredding, and not individual data transfers.

Change-Id: I73cebf8d653cee9c00ce7a952bc75852361fec8b