[a113][pinmux][gauss][gpio] Pinmux Driver for A113
This CL Implements all the functionality required to drive the PinMux on
the aml113.

Future work includes implementing a GPIO driver to drive the GPIO pins and
also a generic mux interface that we can use to generically talk to any
pinmux device.

Eventually we'll want this implementation to sit behind that generic interface.

Change-Id: I4860b31747e6a8cb29c30c12a7a549dfde1f9111