[resolver] Don't force noun output to be Entity references.
This allows the Resolver to output anything that can be encoded as JSON
on the Link. This is a better option for now since it interoperates with
existing mods, and possibly better in the long term (still an open
question).

Also:

* Allows a Daisy to include simple JSON-encoded entity data, and
removes the (non-functional for now) "text" input field. We can add this
back in when we want to do something with it.
* Adds a library for converting an EntityReference into or out of JSON
in a way that can be recognized later as JSON that represents an Entity
reference, or just normal JSON.

MW-276

Change-Id: Ib2af3292cb9895874dfc7c7417649238aca6112c