[ledger] switch cloud sync to use a standalone cloud provider.
This CL switches Ledger to use a standalone cloud provider binary for
cloud operations, instead of relying on baked-in logic.

This CL is bigger than we'd like, but unfortunately it makes limited
sense to switch to the standalone cloud provider only for a subset of
cloud operations - hence here we convert all at once.

To make the already big patch as small as possible, this change does not
do any cleanups that can be left for future patches - including removing
all traces of the network service from ledger/app, moving out the
Firebase and GCS client libraries, etc.

Benchmark results:

baked-in (ToT):

    sync latency (benchmark) ->
      samples 0 to 0: 2346.37137760417ms
      samples 1 to 9: avg 1916.43250472608ms out of 9 samples. (std dev 369.512297926615)
    get and verify backlog (benchmark) -> 2233.77907204861ms

standalone (this patch):

    sync latency (benchmark) ->
      samples 0 to 0: 1979.36031597222ms
      samples 1 to 9: avg 1876.24365113812ms out of 9 samples. (std dev 281.856756470034)
    get and verify backlog (benchmark) -> 1924.14998784722ms

Not sure if the speed-up in sample 0 and backlog is real or just noise,
but we're most likely not getting slower.

LE-318 #done [ledger] switch cloud sync to use a standalone cloud provider.

Change-Id: Ib48fcf4f8a71bb61d1123df9366818935cef2a1a