netconnector: simplify mDNS question/resource send logic
The mDNS implementation was employing a scheme in which 'agents'
could schedule the transmission of questions and records. The
scheduling feature was not being used, except to get specific
record ordering within a message. By employing a much simpler
scheme that doesn't include scheduling, ordering happens naturally
(first in, first out). In addition, there was some code to filter
out redundant records within a message that didn't work (it was
doing pointer comparisons instead of value comparisons). That
code has been dropped...duplications are unlikely in any case.

In addition to just making the code cleaner, this change makes it
much simpler to implement unicast response, which will show up in
a subsequent CL.

Change-Id: I4a9bc32d7e0e8acd922536f2a588709c32a95c12