Media: add SetGain to media client API
This uses the audio_renderer's SetGain interface to set output
gain (volume) on a per-stream basis. Note that GAIN (float db,
with unison == 0.0 and range [-160.0, +20.0] is used here
(aligning with the FIDL interface), although this could easily
be exposed instead as a simple VOLUME multiplier (float volume,
with unison == 1.0 and range [0.00000001, 10.0])

Although clients may not need this immediately, I want to get
this CL out for internal review, so that it is ready to go
when the need arises.

Change-Id: Id10f534fd4b8d65c51c0b5133aaaec76f0d5a620