netconnector: simplify mDNS agents using base class
This CL moves some functionality from the mDNS agent class to the
base class for agents (MdnsAgent). It also replaces the WakeAt/Wake
scheme with the more familiar and versatile PostTaskForTime. This
implementation of PostTaskForTime suppresses callbacks into deleted
agents, so no shared_ptr or weak_ptr scheme is needed.

Change-Id: I616f1d8eb00cc6748df200b7518b0c31ee3526ee