[kernel][vdso/ramdisk] allow the userboot/vdso/ramdisk vmos to free their memory
Previously an extra ref was held on these to make sure they never went
out of scope, since it was not okay to unmap a piece of the BKM.

Once the arm kernel was moved out of the BKM, we can now free the pages
once the vmos go out of scope. For safety purposes, immediately unmap
the pages from kernel space for the userboot and vdso data.

ZX-1237 #done

Change-Id: Ifa181ab63210a4a3b25bbf0901283aec098a8bd5