[fidl][cpp] Don't decode null unions
A union of size zero is considered null, so the tag and value should be
ignored. C++ encodes an invalid handle value when the 0th tag refers to
a handle, but Go and Rust simply zero out those fields. With this change
C++ will skip decoding of a zero-sized union, allowing it to deserialize
a null union with handles.

Change-Id: Icaf840a2300d3aa6d318cbe6d5251d881c3cec56