[x86][processor trace] rename buffers to chunks
A "virtual" processor trace buffer is made up of several pieces,
each piece being an io_buffer_t. This allows having a large buffer
without having to allocate a large chunk of contiguous physical memory;
the trace h/w requires each piece to be contiguous and is physically
addressed. Using the term "buffer" for both the entire "virtual" buffer
and each piece is confusing, so name the pieces "chunks".

This is a public API-breaking change, IPT_API_VERSION bumped to 1.

Change-Id: Ibeb63b589704d3b3f0c59b39bb276d6d2c11afc9