[insntrace] rename buffers to chunks
A "virtual" processor trace buffer is made up of several pieces,
each piece being an io_buffer_t. This allows having a large buffer
without having to allocate a large chunk of contiguous physical memory;
the trace h/w requires each piece to be contiguous and is physically
addressed. Using the term "buffer" for both the entire "virtual" buffer
and each piece is confusing, so name the pieces "chunks".

Use of IPT_API_VERSION is temporary, to walk the patch through zircon pinning.

Change-Id: I46ecebba492d591d7487a09ab53da33b70fa98e8