[ledger] drop on-device repository erase (make the API call a no-op)
This patch makes the EraseRepository() a no-op. In order to erase a
Ledger, one can visit https://fuchsia-ledger.firebaseapp.com .

This change is motivated by the maintenance cost of keeping two
implementations remote erase (which need to orchestrate erase for the
data in the cloud and the data already synced to devices both online and
offline).

We have to keep the dashboard available, as it's the only method of
erase that will work if the current state makes the system crash upon
user startup. (before the on-device erase can be excercised)

LE-332 #comment [ledger] drop on-device repository erase (make the API call a no-op)

Change-Id: I84763c43d0c9039127c07b274e9d910c19631a8a