[modular] drop unneeded repository erase from Ledger test harness
The Reset() logic which erases the Ledger repo after a test should not
actually be needed: the test uses a tmp dir as a repo directory, so all
files will be gone when the test is finished.

We used to have a bug that might have made this dance needed in the past
(LE-304), but AFAIU we should be OK without it now. (Also note that we're
planing to remove the EraseRepository API altogether.)

LE-332 [modular] drop unneeded repository erase from Ledger test harness

Change-Id: I30b3b8bcf3ba0d301328cb3a8b04e3a9e7664d44