[audio][utils] Add Close to AudioDeviceStream
Add the ability to close an audio device stream, resetting it all of
the way to its pre-opened state without destroying it.  Also, be more
rigorous about explicitly unmapping a VMO (if it had been mapped)
before closing the underlying VMO handle.  While it is not strictly
speaking necessary to do this, it is still good style to do so.

Change-Id: I0bb17750cd717d6cd60ca00383b1704c40022b28