[resolver] Implement Resolver v0.1: only verbs
This only performs resolution based on the presence of a 'verb' in the
Daisy and a corresponding 'verb' in the Module's manifest.

Missing is a good fallback for the current JSON Link content, as that is
not passed through. We'll need to decide if we want to have this
fallback or if we try to switch 100% to Entity-based resolution and
Module input.

MW-276

Change-Id: Iebf579609407c91d7410cbe8b30e504a420d532a