[modular] Directly send module state to story controller.
It's confusing to have a fidl connection between objects
in the same process that already call each other directly.

Hence, drop the ModuleWatcher connection between ModuleControllerImpl
and StoryControllerImpl, and let module controller propagate
changes to module state directly to story controller.

This gets rid of some internal bookkeeping data in StoryControllerImpl.

Cleanup the places where module data are collected, created and
accessed in ModuleControllerImpl, ModuleContextImpl, and
StoryControllerImpl: Keep one instance of ModuleData in
StoryControllerImpl::Connection and just hand pointers to it
to ModuleControllerImpl and ModuleContextImpl.

This uncovers some story state transitions that make not much sense:
When a root module is stopped through a module controller obtained from
a story controller, the story transitions to STOPPED, which prevents
new modules from starting. This was previously masked by closing the
module watcher early when stopping a module, but is really as designed.
Hence the design is broken and was saved until now by a broken
implementation. For now, we change the update_story test to not
stop the root module so as to prevent the spurious transition to
the STOPPED state. Presumably, only StoryController.Stop() should
transition the story to STOPPED and nothing else, specifically not
any module stopping. Likewise, when the root module calls itself Done(),
the story controller stops it and also transitions the story state to
DONE, with the same consequences for newly added modules. More discussion
in FW-344.

FW-333 #done
FW-344 #comment Marked with TODO meaningless Story state transitions.

Change-Id: If0f64b3fac9a8d99175266cec36d6b67add9cc2d