[x86][hypervisor] Fix present-bit in page fault.
According to the Intel SDM, bits 3 to 5 of the EPT violation exit
qualification indicate whether the PTE marked the page as read, write,
or execute, respectively. There is no bit to indicate whether the PTE
was present.

Since we now require that all traps be set explicitly, we can
additionally early-exit if there was no trap associated with the access
and it is outside guest physical address space.

Change-Id: If2c2ec93b6972f7b03b9667518cf76878c89a1fe