netconnector(mDNS): postpone distribution of host name
This CL changes the mDNS 'agents' and transceivers so that they can
defer getting a host name until relatively late. This is to support
probing for A/AAAA resource conflicts, which may require the host
name to be changed.

In the cast of agents, the host name is no longer provided in the
constructor. Instead, the host name is specified when the agent is
started. This allows agents to be instantiated before the host name
is known. Most agents aren't started until the host name is
determined (the exception will be the A/AAAA prober).

In the case of transceivers, the host name is no longer provided in
Start. The transceivers can function without it (as long as they
aren't asked to send any A/AAAA records), and they allow the host
name to be set multiple times.

Change-Id: Ieded89623453d7a94d6016932567c2399a754879