[zircon][abi] limit wait_many to eight handles
For larger wait sets, ports are the correct tool for the job.
Wait Many is intended to cover the "I need to wait for 2 or 3 things"
case.

Introduced the ZX_WAIT_MANY_MAX_ITEMS constant and updated the manual page.

This avoids heap allocations in the kernel for wait many (which could
be as large as 50k with the old 1024 handle limit), and will allow us
to move to an always-inline strategy.

It discourages users of wait many from using large sets which, even
ignoring the allocation issue, are highly inefficient (having to create
and register a wait state observer for every handle and deregister
after).

This will impact poll(), ppoll(), and select() which will fail for
more than eight items now.  Slow-path versions could be provided,
but better would be to reintroduce epoll() on top of ports.

Change-Id: Ib8874a8880ac285441789c5f69abaf00a92976cb