netconnector(mdns): change the way subtypes are announced
This CL changes the mdns FIDL definition to align subtype support
with the way certain devices want to behave.

Previously, the subtypes to be announced for a given instance were
specified once when the instance is published (via AddResponder).
However, the devices in question want to be able to announce subtypes
asynchronously with respect to publishing the root instance. The set
of subtypes announced for such a device changes over time. Now
AddResponder no longer accepts a list of announced subtypes. Instead,
a SetSubtypes method allows the set of subtypes associated with an
instance to be changed at any time.

Previously, the mDNS implementation announced subtypes by sending
a complete set of resource records (PTR, SRV, TXT, A, AAAA) for each
subtype. Now that set is only sent for the root instance, and subtype
PTR records are issued seperately.

The devices in question also need to be able to change their TXT
resource record at any time. The new ReannounceInstance method
supports this by causing the instance to reannounced, giving
clients who have used AddResponder an opportunity to issue an
updated TXT record.

Change-Id: I03c2e348444443b033c1f0205bb1c406786a7a6a