media: Fix DCHECK on reconnect
This CL fixes issues with disconnecting and reconnecting a producer.
Specifically, the SharedBufferSetAllocator for a producer needs to
be reset before it can be reused for a new consumer. We do this on
Connect rather than Disconnect so that the allocator can continue to
function for deallocations after Disconnect. A guard has also been
added in case deallocators occur after Reset. There is no memory leak
implied, because the vmo in question has been released.

Change-Id: I5a7ed129467df7eb15cecf7d2905eaf1cd7bb586