[blobstore][host] Fix bug which attempts to read from beyond EOF
Bug: When a file is added to Blobstore, it is mmapped, and
we read from that mmapped file in block-sized chunks when
writing to the Blobstore partition. If this file's size is
not a multiple of the blobstore block size, then we may
attempt to read beyond the end of the mmapped file (up to the
nearest block size, which is NOT always the same as the nearest
page size).

Fix: If we realize that reading from the file (in block-sized chunks)
goes beyond EOF, copy to a temporary buffer, and read from there
instead.

Change-Id: I92c793f04c83b717dd8b2fdb67cd7ea42a137e4c