[libc] Initial libc module map
Adds a clang module map to libc.

Eventually this map will be consumed by clang to:
- improve build times by allowing the compiler to cache AST passes of
headers in the system to use over again for each build unit instead of
reparsing them each time (similar to a PCH).
- improve compile-time scalability and ambiguitiy by being able to
specify link dependencies for each header.
- allow symbol access in other tools like lldb by giving tools access
to serialized AST representations of the code.
- remove header fragility since each library is parsed as a standalone
enity. This eliminates the need to use defensive tricks with symbol
naming by not allowing one library to affect how another library is
compiled and emits its symobls. This also eliminates include-order
dependencies with headers.

Depending on the direction that module spec goes in the C++ Module TS
for C++17, this file may or may not be used other C\C++ compilers in
the future.

This module map is also used by Swift to import platform headers.

Change-Id: I2ab302d7b37e8e7aad7fe497a6ebf4e633655bc8