[ledger] Move in-repo third_party code to peridot root.
For context, we applied for new repos to move these bits out of repo but
were advised not to. As the second best thing, this patch moves them to
root, allowing these and future instances of in-repo third_party code to
be excluded from linters/formatters using a simpler filter.

Change-Id: I0ad2892e5ee531d354813ab70f61c08a77afe478