Adopt one clang-tidy config for the perdiot repo.
This patch converges on a single .clang-tidy config derived from the
latest one used in Ledger (and explained in lint.md) that knows how to
filter out warnings in the generated FIDL files and in-repo third_party
code.

After this patch, one can run:

`git file-tidy --all --out-dir out/debug-x86-64`

in any place in the the repo, and get a list of issues to fix.

Change-Id: Ia3c5a30e3dbcd9a0b123327f176b6714c40401e1