netconnector(mdns): prevent empty messages from being sent
This CL fixes a bug that causes empty messages to be sent. When an
expiration was published to agents via SendResource, a message was
sometimes created. Typically, no questions or resources were added
to that message, so an empty message would be sent. This CL checks
for the expiration case before finding/creating a message, so empty
messages are no longer sent.

Change-Id: I05197cf226484713ac43882c4981bbee99e90569