[audio] Fix filter selection.
Right now, the auto-magical filter selector is supposed to select the
point resampler only if the source sample rate is an integer multiple
of the destination sample rate (eg; source is 96k dest is 48k), but in
reality it was doing the opposite.

Fix the code to bring it into line with the comment.

Change-Id: Ibc9afe56fe5006090ef552db3b1ec0630e5a4ef3