[fx] Don't start goma if it is already running
Running "gomacc port" is a much faster way to check if goma is running than
asking goma_ctl. This CL uses this check to be faster in the common case when
goma is already running.

We might want to move this check to "fx build", but for now it's easier to
implement the check in "fx set".

Change-Id: I43043588b41a9219692c738d335b4074fecb8feb