[docs] Use long-form verb names in docs.
There's been some confusion around what a short-form verb name (like
"Navigate") implies about the role of verbs. This explicitly uses
long-form names in examples and docs.

The format is up for discussion in the future. For now, this is fine,
and we can change our minds once we have more information.

This doesn't change the examples for Entity types, and this may be worth
doing as well.

MW-309 #done

Change-Id: I4d9863349789df1b53be0ea92ea387a23f903556