[device][intel-pm] more sample frequencies, split fixed counters up
Until we have the ability to set arbitrary sample rates, it's help to
have a bit more flexibility.
The fixed counters can take up a lot of space in the trace buffer,
but they are useful: Allow specifying a subset of them.

This patch reapplies commit bf84825199d9b3347cea17628b2bfbf8003d9565

Change-Id: Ie4873ffc2cf6a5a34199a1b2f4a5c9a15d16c92f