[maxwell][suggestion engine] Remove remaining 1x1 suggestion updates
1x1 were used in the V0 suggestion API but are being phased out in favor
of mass invalidation. Work done on suggestion engine so far has removed
about half of the 1x1 updates but missed many, resulting in duplicate
events being sent for query, the first of which had 0 confidence values,
which were throwing off Dugong's jackpot behavior.

This change gets rid of the remaining 1x1 updates in all but the
interruption channels, which may not end up being migrated to mass
invalidation.

Change-Id: I4e855d41a9b70c1f43e740a6e4b5af55ae8fcda8