[wlan][supplicant] Introduce Supplicant
In order to better support other handshakes, such as Group Key ones, the
EAPOL state machine should be changed a little.

First, a new layer "Supplicant" is introduced. The Supplicant is aware
of the overall system state, e.g., it controls KEK, KCK and Replay
Counter. This is necessary to decrypt encrypted messages of not just the
Four-Way Handshake but also others, such as the already mentioned Group
Key one.
The Supplicant is also a demuxing entity for the various handshake
types. E.g., it understands whether an EAPOL Key frame is part of a
Four-Way, Group Key or Peer Handshake and forwards them accordingly; or
drops the frames if they are not allowed in the current system state.

To keep the CLs small and refactoring slim, I first introduce the
Supplicant layer which only forwards frames to the Four Way Handshake.
A follow-up CL will move Supplicant specific state knowledge, such as
KEK, KCK etc., into the Supplicant and provides utility methods to
access these from within the active Handshake.
Once this change landed, the integrity and state verification is moved
into the Supplicant and only Handshake specific checks will reside
inside the Handshake implementation. This (1) adds additional guarantees
to each handshake about the Key frame, (2) reduces complexity for
verifying correctness of incoming frames, (3) more secure because we
drop earlier.

Change-Id: Icbea4b5f4eb72501128030692bcef17fe2d3a8ea