Use symlinks for host_toolchain variant binaries
When "copying" a host executable selected to build as a variant,
use a symlink instead of a hard link.  This ensures that the
dynamic linker's executable-relative logic for finding shared
libraries will find the variant builds of the shared libraries
instead of the main host toolchain's builds.

TO-558 #comment Made sure ASan builds of host tools use ASan shlib builds.

Change-Id: I718b2dce62633276c55395a503076af40ffa8f2c