[trace] drop support for on-device dashboard upload
We decided to put the dashboard upload logic in the recipe, rather than
on-device. This patch removes the dashboard upload logic from the trace
tool. In the next patch, we will add support for exporting the benchmark
results to a JSON file, which a recipe would pull down and parse.

US-370 #comment [trace] drop support for on-device dashboard upload

Change-Id: I6208a4ed46363eea84cc5737475a4a7f5aba2b36