Prepare package() template for the sandboxening
Currently, the package() template defaults to placing its contents in
the system image. This CL is the first step towards switching the
default to placing the contents of the package inside an actual package.
(For now, we're still using raw FARs, but we'll eventually switch to
real packages.)

This CL introduces the |system_image| argument to the package() template,
which is the opposite of the |archive| argument. The idea is to migrate
all existing non-archived packages to setting |system_image = true| and
then flip the default to false.

This CL also introduces the |binary| argument, which is an easier way of
putting a binary in the |bin/app| path inside the package (which is the
default path executed by appmgr when running the package).

Change-Id: Ic6e7e8e48cab0419608ffa249344a1af38771808