[netstack] Add a NotificationListener FIDL interface.
This defines a single function, OnInterfacesChanged, which is notified
with the list of interfaces whenever the state of interfaces changes.

Also added support to listen for eth status changes from zircon, which
then trigger an OnInterfaceChanged notification.

Also removed getaddrtest and consolidated it into a
netstack_service_test, which includes a test of OnInterfacesChanged.

NET-134 #done
NET-189 #progress

Change-Id: I440e1c8f8ad68523e24d0850d144ff25f6e405d7