[zicon][kernel] more RefPtr<> return by const&
This change follows Jeff's change in [1] which
avoids atomic ref up + atomic ref down in

    object.inner()->Operation()

When inner() returns a RefPtr<T> temporary.

While being "pretty safe" for the other normal
cases. See discussion in [1] for more detail. The
only practical dangerous case is

auto& inner = object.inner();

Which we don't do and is easy to spot.

[1] https://fuchsia-review.googlesource.com/c/zircon/+/79676

Change-Id: I38f75a1125c291c653534ee91a2096951ecf06ea