[maxwell] Unify QueryContext and UserInput
The suggestion ranking changes introduced a query context structure
that overlapped the UserInput FIDL struct. This CL will unify those
two structures with only UserInput remaining. This removes some
duplication and will allow the ranking features to be aware of
features of the query that are not currently passed from SysUI.

Additional features:
* Adds query timeouts
* Fixes missing processing changed/idle status on early suggestion
  selection
* Pins kronk-2017.11.08.10.55

Change-Id: I58e01eeef8192e94b47b5659a8cc36a4eb9709a0