[build] Add the system_image property to packages
Currently, packages defined with the package() template are put in the system
image by defaul. We're working to change the default to separating packages
from the system image. In order to make a smooth transition, this CL whitelists
all the existing packages to keep their current behavior. Over time, we will
likely migrate many of these packages out of the system image.

Change-Id: Iee913f5f1706b711002826acefc3e12f2895dc47