[audio] Fix a leak of renderer->output links
Fix a leak which crept in with CID
I71e105c9327e63849dbd466b49e4f6134d351d84.  Here is the basic idea...

Renderer and Outputs are linked together using a structure which is
appropriately named a RendererToOutputLink.  Outputs maintain a set of
links to renderers, and renderers a set of links to outputs.

Currently, when a renderer goes away, it does not go over the list of
output links it is currently bound to and attempt to remove the links
from the outputs' set.  Doing so would require it to hold the output's
link set lock, something it does not want to do for a number of
reasons (details available on request).  Instead, it was relying on a
lazy cleanup approach implemented by the output itself.  The idea is
that the link has a weak pointer back to the renderer.  When an output
picks up the link for a renderer, it would promote the weak pointer to
a strong pointer, and if that promotion failed, it would remove the
link from its set of links, because it knows that the renderer has
gone away.

During this recent re-factor, this attempt to promote went away.  This
means that renderers would go away, but their links would never get
cleaned up.  There would never be anything to render present in the
links, but the output would still be tracking a pointless link.

All of this junk needs to get re-worked for a number of reasons.  Some
of this work is in flight and some is tracked by other bugs.  For now,
however, I've added back in the attempt to promote/cleanup-on-failure
behavior, so at least we are not leaking no-op links.

Change-Id: I134da7b6599411ebb17fabecfcdd3772107901e4