Support building multiple zircon projects at once and always build qemu-arm64 when targeting arm64
This allows passing multiple comma-delimited projects to
build-zircon.sh that will all be built. It also makes sure that the
zircon-qemu-arm64 project is always built when an arm64 target so the
rest of the build can use the sysroot from the qemu build. In the
future, we should build only the target-specific parts of Zircon for
other arm64 targets and teach the boot server and installer parts of
the system how to find these parts instead of pointing the entire build
at a different Zircon build for each arm64 target.

Change-Id: Ifc4ed662bb6636714169c1f3ee516577244ff3fe