[bt][common] Optionally serialize 32-bit UUIDs as 128-bit
Added an argument to UUID::Copy and UUID:CompactSize to optionally
fall back to a 128-bit representation when the UUID is 32-bit. This is
useful for GATT which does not allow 32-bit UUIDs for the attribute
types that it defines.

Also added UUID::CompactView which returns a BufferView without
copying.

NET-179

Change-Id: I9b7f18e2a81f5e5357ba6075d57daf03799bcdf8