[x86][hypervisor] Inject interrupts on VCPU resume.
If there are interrupts pending, inject them at VCPU resume instead of
waiting for an EOI. This also corrects the situation where we are
executing on the same CPU as the VCPU when we issue an interrupt.
Without this, the interrupt processing would be delayed until an EOI was
seen.

ZX-1340 #done

Change-Id: I4dabbc616a8750190a424948982b0e1b0b911b5e