[installer] Don't spell x86-64 incorrectly
Previously, we were trying to tell build-zircon to build a specific target, but
we weren't spelling the target name correctly, which caused the script to error
out. Now we don't say which target we want because the default is x86-64 and
that's the only arch on which the installer works anyway.

Change-Id: Ib1d3096c242c552b869f46aa27b16d1bac141e80