[ledger] Fix coroutines.
The issue was happening when a coroutine was continued inside the
destructor of the main method of another coroutine. The coroutine was
then yielded inside the erase method of the vector containing all
coroutine handler which was left in a non-coherent state.

This CL fix this problem by running the destructor of the main method of
a coroutine inside the coroutine itself.

LE-349 #done

Change-Id: I74ea7319f4abff99ac6eee6264ebe0584af492a4