[devmgr] Reconsider fallback drivers after loading /system
When devmgr.require-system is set to true, the device manager will
ignore any drivers marked as fallback drivers before the /system
partition is loaded and drivers in that location are added. This
modifies the manager to re-consider these drivers once /system is
scanned so the fallback drivers can still be matched if there are no
better matches in /system.

Change-Id: I0b516cfd8e96fffd8b44f2602d713a34e875f6da