[media][utils] Make TimelineFunction return rate as reference
There is really no reason for TimelineFunction to be forcing copy
construction of a TimelineRate every time you ask for a rate.  Change
the inline access to return a const& so that the compiler will either
just access the structure directly, or make a copy, depending on how
the user chooses to use the rate() function.

Change-Id: I0607154d49506aa99a65cfb59650c90b9322f7c1