[audio][mixer] AudioLink refactor
Major refactor of how links work system wide.  This lays the
groundwork for being able to link the various audio object types in
the system to each other, and for storing the per link state for links
which source from packet sources vs. links which source from ring
buffer sources.

Change-Id: I8792da0c7ee6082034c9a43c081b0a2f67d27fc0